Added MimeType checking
This adds support for the `mimeType()` function on `String?` types. The
previous implementation allowed to pass nil values to the internal
`MimeType(ext: String?) -> String` function but `.mimeType()` could
never be called on a nil String because that was an extension on the
`String` type.
This allows us to do this
```swift
let testString :String? = nil
testString.mimeType() //  -> "application/octet-stream"
```
Instead of this
```swift
let testString : String? = nil
(testString ?? "").mimeType() // -> "application/octet-stream"
```
Or worse
```swift
let testString : String? = nil
testString.extension?.mimeType ?? "application/octet-stream" //
->"application/octet-stream"
```
This new implementation is used in the router and fixes #25