Added MimeType checking
This adds support for the `mimeType()` function on `String?` types. The
previous implementation allowed to pass nil values to the internal
`MimeType(ext: String?) -> String` function but `.mimeType()` could
never be called on a nil String because that was an extension on the
`String` type.
This allows us to do this
```swift
let testString :String? = nil
testString.mimeType() //  -> "application/octet-stream"
```
Instead of this
```swift
let testString : String? = nil
(testString ?? "").mimeType() // -> "application/octet-stream"
```
Or worse
```swift
let testString : String? = nil
testString.extension?.mimeType ?? "application/octet-stream" //
->"application/octet-stream"
```
This new implementation is used in the router and fixes #25

Forgot semicolon in Router

Changed extension to ``extension``

Updated the Optional extension to allow any kind of optional. Thought it would be unsafe to force cast an Optional of a different type to a String? but turns out to work just fine

Added safety to the optional extension. Casting a non nil Optional type of any other type than String would previously cause a crash.

Oops, forgot parentheses

Updated the Router to use optional chaining and removed the optional extension on MimeType

Added a static test javascript file and updated the tests to check for the mime type served by the router for static files.